Removed file read warning.

+++ src/Util.cpp
@@ -10,7 +10,7 @@ namespace util
 		if (in)
 		{
 			in.seekg(0, std::ios::end);
-			contents.resize(in.tellg());
+			contents.resize(static_cast<unsigned int>(in.tellg()));
 			in.seekg(0, std::ios::beg);
 			in.read(&contents[0], contents.size());
 			in.close();
